Adapt GPIO pins to match Agat PCB by "bleep lab"
The controller GPIO pins are taken from the firmware for the 12x2 display - in the OLED version of
the firmware they for some reason have changed them.

ATTENTION: After the commit, the buttons are located differently to how they are labeled on the
display: the display shows help icons at the bottom in the order "A  B  C", and the buttons on the
PCB are located in the order "B  C A".